Media window popup (#48)
* Window popup for media (#43)

* Added window popup for images

Opens photo attachments from tweets in a new window instead of a tab in
your current browser.

* Re-indenting, casing modified based on comments and debug stuff removed

* Forgot to comment out dev window.

* Added ImagePopup.html, added imagePopup.scss, and polished the window

* Images now scale proportionally when too big

* Re-added changes from another pull request which fro some reason wasn't added here

Additionally I also made it possible to see thumbnails of video and
gifs..

* Fixed an bug where after loading an image the second time, win would be null resulting in TypeError

* Added window popup for images

Opens photo attachments from tweets in a new window instead of a tab in
your current browser.

* Re-indenting, casing modified based on comments and debug stuff removed

* Forgot to comment out dev window.

* Added ImagePopup.html, added imagePopup.scss, and polished the window

* Images now scale proportionally when too big

* Re-added changes from another pull request which fro some reason wasn't added here

Additionally I also made it possible to see thumbnails of video and
gifs..

* Fixed an bug where after loading an image the second time, win would be null resulting in TypeError

* Window displays mp4 links now in addition to images. Renamed imagePopup to mediaPopup. Moved resize into it's own method

Although that has no real use other than shortening the code. it’s
still enclosed to the JavascriptExecution but if it’s needed for other
things in the future, I think we can move it somewhere else.

* Fixed indentations, added newLine space to html file

* Put newline at end of file

* Unify indentation

* Unify filename formats

All filenames are chain-case.
While I want it to be snake_case, using chain-case as JS for now.

* Drop trailing whitespace

* Fix inconsistent indentation

* Cosmetic changes

* Use arrow function

* This line is unused

* This line is unused too

* Reduce lines of code

* Drop unnecessary space before equal

* Unify variable names

* Don't check loaded

First, you must not use duplicated id attribute.
Second, clearInterval's argument is undefined.

Removing it for now since it may not work.

* Use spaces in executed JavaScript

* Pass variables by ipc

* Fixed window sizing, removed win.center, scaled approximately from center now too

* Fix my fatal mistake in 22715bb51e2af79b845211075cbe88b74fb3c70a

* Drop trailing whitespace

* Don't executeJavaScript

* Cosmetic changes

* Fix to require BrowserWindow and ipcMain in function

* Add FIXMEs

* Video fix  (#49)

* fixed video not working after new commits

* Shouldn't be done to img for some reason...

* Fix video resizing problem

* Added hls support for media window (#50)

* Added hls support for media window, dependency for hls.js added to package

* Cosmetic changes

FIXME: use import has been refactored for ES6, changed name of
video.title for convenience and other small cosmetic changes

* use once instead of on to avoid memory leaking listeners

There is no need for us to set up a new listener for each time we open
a window.
FIXME: refactored ipcRenderer to ES6, but was that all?

* More cosmetic changes, made it possible to drag window

Also between this commit and the last, which involved using ‘once’
instead of ‘on’ fixed the issue that previous windows would be resized
to the new window size.

* Padding should not be added to other platforms, height should always add 20

* Removed spaces, aded semicolon to various lines.

And other adjustments.